FLEX AND BISON INTERACTION (BUG) This commit focus its attention on the compiling mechanisms between flex and bison. Moreover it adds a great part of syntax analysis. Important commit are 5,6,8
1)DOCUMENTATION: adds some documentation throughout all the project.
Creates the folder "verbatim". It is the location where all the files
included by the command \verbatiminclude should go. It is specified in
the doxygen variable EXAMPLE_PATH.
2) CONCEPTS: changes the abstract syntax tree idea generated by bison.
The tree generated is built with the idea expressed inside the file
"ideas on syntax tree". "makefile explained" now explains more deeply
the interaction between bison and flex.
3) OLD FILES: creates a directory called "oldfiles": it contains the ex
token-definitions.h manually created. Now it is useless because bison
automatically generates it. The file has been moved in oldfiles because
of its code is well organized.
4) RUN CONFIGURATIONS: added the run configuration of the syntax_test.c
5) LEXICAL: removes lex-definitions.h: it is now automatically created
by bison. lexval variable is now called "lexVal". Adds the yylineno
directive: now to know where the lexical token just read from the file
is located inside the file itself you can use the varialbe "yylineno".
To use it ypou have to declare it as such:
	extern int yylineno;
and then use it. Since bison thought that token like '(' and
TK_OPENPARENTHESIS were different tokens the commit deletes tokens like
TK_OPENPARENTHESIS (actually all the sugar tokens are deleted). For
token composed by only a character (sugars and >,<,+,-,*,/) flex return
yytext[0].
6) SYNTAX: adds in Node some function: buildSyntaxTreeGraph for the
drawing of the abstract syntax tree; it should be tested though. In
NodeTyype is added function to print the NodeType string and the
functions isTerminal/isNonTerminal. Now syntax.y:
 - adds yyerror function: without it syntax.c cannot be compiled
 - defines INITIDNODE, a macro introduced to improve code readability
 - removes tokens indicating sugars like TK_OPENPARENTEHSIS
 - adds start buson directive to improve code readability
 - rules that does not use the "$$=initNode();" code are now preceded by
the comment /*Does not create a node*/: this is useful because that node
must be created by the parent node. These derivation are practically
only lists.
 - fills the construction of all the syntax tree. It should be tested
though.
To ensure the correct working of syntax.y syntax-test.c has been
created, though it is only a STUB. The stub works with a test sol file,
named "sol01.sol".
7) PROJECT: adds incude directiv einside C/C++ -> General -> Paths and
Symbols -> Include paths. Adds some Make targetentry like "Syntax test"
or "Lexical test". ".gitignore" now ignore the grp directory and the
files generated by flex and bison.
8) MAKEFILE adds -std=c99 (but not -pendantic) in makefile as suggested
http://stackoverflow.com/questions/1783792/why-does-gcc-report-implicit-declaration-of-function-round.
No improvements are foound though. Adds the ouputs of flex and bison in
clean rule. The header file generated by bison is now called
"token-definitions.h", not syntax.h: this because the name "syntax.h"
wsas not explicit enough to understand that that file contained a list
of lexical tokens. To do so check-dirname had to be included in the
prerequisites.

********************** BUG *****************************
1) in flex "a+3" is detected as id (a) and intconstant (+3): it should
be id (a), plus sign (+) and intconstant (3) instead. This should be
done using the rules precedence in flex by effectively using the maximal
munch idea.